tests/wpt/tests/xhr/open-url-base-inserted-after-open.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: open() resolving URLs - insert &lt;base> after open()</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#the-open()-method" data-tested-assertations="following::ol/li[2]/ol/li[2] following::ol/li[7] following::ol/li[14]/ul/li[2]" />
    <link rel="help" href="https://xhr.spec.whatwg.org/#the-responsetext-attribute" data-tested-assertations="following::ol/li[4]" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        var client = new XMLHttpRequest(),
            base = document.createElement("base")
        base.href = location.href.replace(/\/[^/]*$/, '') + "/resources/"
        client.open("GET", "folder.txt", false)
        document.getElementsByTagName("head")[0].appendChild(base)
        client.send(null)
        assert_equals(client.responseText, "top\n")
      })
    </script>
  </body>
</html>
